+++ comics/comics/wumovg.py
@@ -1,4 +1,4 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import HeltNormaltCrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -9,12 +9,9 @@ class ComicData(ComicDataBase):
     rights = 'Mikael Wulff & Anders Morgenthaler'
 
 
-class Crawler(CrawlerBase):
+class Crawler(HeltNormaltCrawlerBase):
     history_capable_date = '2013-01-26'
     schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
-    time_zone = 'Europe/Oslo'
 
     def crawl(self, pub_date):
-        url = 'http://heltnormalt.no/img/wumo/%s.jpg' % (
-            pub_date.strftime('%Y/%m/%d'))
-        return CrawlerImage(url)
+        return self.crawl_helper('wumo', pub_date)